T2/requirements.txt
bleach==1.5.0
enum34==1.1.6
html5lib==0.9999999
Markdown==2.6.9
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
protobuf==3.4.0
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc2
Werkzeug==0.12.2
